Split the TxPool interface and its implementations into separate files
